Template For Recipe Pages (apple pie)
This is the template we will be using to create all of the recipe pages, this has the place for the images, ingredients, and instructions. Right now its is all filled in form the apple pie recipe. But we can just fill in the other information later. This is missing most of the css.